Consolidated directory of vault scripts and contracts
